moved TestExporterService to its own file
